% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_pof.R
\name{baixar_pof}
\alias{baixar_pof}
\title{Baixar e descompactar microdados da POF}
\usage{
baixar_pof(pasta_destino)
}
\arguments{
\item{pasta_destino}{Pasta na qual os microdados serão guardados (se não existir, será criada automaticamente)}
}
\value{
Os microdados na pasta designada
}
\description{
Baixar e descompactar microdados da POF
}
\examples{
baixar_pof("./microdados");
}
