% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_despesa_media.R
\name{calcular_despesa_media}
\alias{calcular_despesa_media}
\title{Calcular médias mensais de despesa (por tipo de despesa)}
\usage{
calcular_despesa_media(
  df_pof_despesa_base,
  tipo_despesa = 0,
  df_pof_morador_base,
  uf = "all",
  regiao = "all"
)
}
\arguments{
\item{df_pof_despesa_base}{O nome (string) do dataframe com os dados de despesa. Ver ler_pof_despesa}

\item{df_pof_morador_base}{O nome (string) do dataframe com o registro MORADOR. Ver ler_pof_geral}

\item{tipo_despesa=0}{Tipo (ou tipos) de despesa. Ver indice_despesa}
}
\value{
O valor médio real mensal do tipo de despesa escolhido
}
\description{
Calcular médias mensais de despesa (por tipo de despesa)
}
\examples{
calcular_despesa_media(df_pof_despesa_base = "pof_despesa", tipo_despesa = c(0,1,2), df_pof_morador_base = "pof_morador");
}
\seealso{
ler_pof_despesa, ler_pof_despesa_todas, montar_tabela_despesa_uc
}
