% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_rendimento_medio.R
\name{calcular_rendimento_medio}
\alias{calcular_rendimento_medio}
\title{Calcular médias mensais de rendimento (por tipo de rendimento)}
\usage{
calcular_rendimento_medio(
  df_pof_rendimento_base,
  tipo_rendimento = 0,
  df_pof_morador_base
)
}
\arguments{
\item{df_pof_rendimento_base}{O nome (string) do dataframe com os dados de rendimento. Ver ler_pof_rendimento.}

\item{df_pof_morador_base}{O nome (string) do dataframe com o registro MORADOR. Ver ler_pof_geral}

\item{tipo_rendimento=0}{Tipo (ou tipos) de rendimento. Ver indice_rendimento}
}
\value{
O valor médio real mensal do tipo de rendimento escolhido
}
\description{
Calcular médias mensais de rendimento (por tipo de rendimento)
}
\examples{
calcular_rendimento_medio(df_pof_rendimento_base = "pof_rendimento",
tipo_rendimento = c(0,1,2),
df_pof_morador_base = "pof_morador");
}
\seealso{
ler_pof_rendimento, ler_pof_rendimento_todas, montar_tabela_rendimento_uc
}
