% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_valor_despesa_mensal_uc.R
\name{calcular_valor_despesa_mensal_uc}
\alias{calcular_valor_despesa_mensal_uc}
\title{Mean monthly expenditure values (by type of expenditure)}
\usage{
calcular_valor_despesa_mensal_uc(
  pof_despesa = 0,
  tipo_despesa = 0,
  uf = "all",
  regiao = "all"
)
}
\arguments{
\item{pof_despesa}{The name of the df with the expenditure data (string). Default to load a new df.}

\item{tipo_despesa=0}{The type (or types) of expenditure. Default to total expenditure.}

\item{uf="all"}{The relevant federal unit (numeric)}

\item{regiao="all"}{The relevant macroregion (character code)}
}
\value{
The mean deflated monthly expenditure
}
\description{
Mean monthly expenditure values (by type of expenditure)
}
\examples{
calcular_valor_despesa_mensal_uc(tipo_despesa = c(0,1,2));
calcular_valor_despesa_mensal_uc(pof_despesa = "df_expenditure", tipo_despesa = c(0,1,2), regiao = "N");
}
