% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_valor_rendimento_mensal_uc.R
\name{calcular_valor_rendimento_mensal_uc}
\alias{calcular_valor_rendimento_mensal_uc}
\title{Mean monthly income values (by type of income)}
\usage{
calcular_valor_rendimento_mensal_uc(
  pof_rendimento,
  tipo_rendimento = 0,
  pof_morador,
  uf = "all",
  regiao = "all"
)
}
\arguments{
\item{pof_rendimento}{The name of the df with the income data (string). See ler_pof_rendimento.}

\item{pof_morador}{The path to the microdata folder}

\item{tipo_rendimento=0}{The type (or types) of income. Default to total income. See indice_rendimento}

\item{uf="all"}{The relevant federal unit (numeric). NOT IMPLEMENTED YET}

\item{regiao="all"}{The relevant macroregion (character code). NOT IMPLEMENTED YET}
}
\value{
The mean deflated monthly income
}
\description{
Mean monthly income values (by type of income)
}
\examples{
calcular_valor_rendimento_mensal_uc(pof_rendimento = "df_income", tipo_despesa = c(0,1,2), path_midrodata = "./microdata_folder");
}
