% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indice_despesa}
\alias{indice_despesa}
\title{Dictionary with codes of expenditure and its respective categories}
\format{
A data frame with 93 rows and 2 variables:
\describe{
\item{NIVEL}{numeric, code used as parameter}
\item{DESCRICAO}{Description of expenditure category}
}
}
\usage{
indice_despesa
}
\description{
Dictionary with codes of expenditure and its respective categories
}
\keyword{datasets}
