% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indice_rendimento}
\alias{indice_rendimento}
\title{Dictionary with codes of income and its respective categories}
\format{
A data frame with 17 rows and 2 variables:
\describe{
\item{NIVEL}{numeric, code used as parameter}
\item{descricao}{Description of income category}
}
}
\usage{
indice_rendimento
}
\description{
Dictionary with codes of income and its respective categories
}
\keyword{datasets}
