% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_pof_despesa.R
\name{ler_pof_despesa}
\alias{ler_pof_despesa}
\title{Load POF microdata - expenditure information only}
\usage{
ler_pof_despesa(microdata_file)
}
\arguments{
\item{microdata_file}{The specific microdata file to be loaded}
}
\value{
The file-specific microdata as a dataframe (all cols to character)
}
\description{
Load POF microdata - expenditure information only
}
\examples{
ler_pof_despesa("./data/microdatafile.txt");
}
