% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_pof_despesa_todas.R
\name{ler_pof_despesa_todas}
\alias{ler_pof_despesa_todas}
\title{Load POF microdata - all expenditure information}
\usage{
ler_pof_despesa_todas(microdata_folder)
}
\arguments{
\item{microdata_folder}{The path to the FOLDER that contains the microdata txt files}
}
\value{
All expenditure information available
}
\description{
Load POF microdata - expenditure information only
}
\examples{
ler_pof_despesa_todas ("./data_folder");
}
