% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_pof_geral.R
\name{ler_pof_geral}
\alias{ler_pof_geral}
\title{Load POF microdata}
\usage{
ler_pof_geral(microdata_file)
}
\arguments{
\item{microdata_file}{The specific microdata file to be loaded}
}
\value{
The file-specific microdata as a dataframe (all cols to character)
}
\description{
Load POF microdata
}
\examples{
ler_pof_geral("./data/microdatafile.txt");
}
