% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_pof_rendimento.R
\name{ler_pof_rendimento}
\alias{ler_pof_rendimento}
\title{Load POF microdata - income information only}
\usage{
ler_pof_rendimento(microdata_file)
}
\arguments{
\item{microdata_file}{The specific microdata file to be loaded}
}
\value{
The file-specific microdata as a dataframe (all cols to character)
}
\description{
Load POF microdata - income information only
}
\examples{
ler_pof_rendimento("./data/microdatafile.txt");
}
