% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_pof_rendimento_todas.R
\name{ler_pof_rendimento_todas}
\alias{ler_pof_rendimento_todas}
\title{Load POF microdata - all income information}
\usage{
ler_pof_rendimento_todas(microdata_folder)
}
\arguments{
\item{microdata_folder}{The path to the FOLDER that contains the microdata txt files}
}
\value{
All income information available (as a dataframe)
}
\description{
Load POF microdata - income information only
}
\examples{
ler_pof_rendimento_todas ("./data_folder");
}
