% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montar_tabela_despesa_uc.R
\name{montar_tabela_despesa_uc}
\alias{montar_tabela_despesa_uc}
\title{Calcular despesa mensal de cada UC (por tipo de despesa)}
\usage{
montar_tabela_despesa_uc(
  df_pof_despesa_base,
  tipo_despesa = 0,
  df_pof_morador_base
)
}
\arguments{
\item{df_pof_despesa_base}{O nome (string) do dataframe com os dados de despesa. Ver ler_pof_despesa}

\item{df_pof_morador_base}{O nome (string) do dataframe com o registro MORADOR. Ver ler_pof_geral}

\item{tipo_despesa=0}{Tipo (ou tipos) de despesa. Ver indice_despesa}
}
\value{
Um dataframe com todas as UC (registro MORADOR) com os valores de cada tipo de despesa como colunas
}
\description{
Calcular despesa mensal de cada UC (por tipo de despesa)
}
\examples{
montar_tabela_despesa_uc(df_pof_despesa_base = "pof_despesa", tipo_despesa = c(0,1,2), df_pof_morador_base = "pof_morador");
}
\seealso{
ler_pof_despesa, ler_pof_despesa_todas, calcular_valor_despesa_mensal_uc
}
