% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montar_tabela_rendimento_uc.R
\name{montar_tabela_rendimento_uc}
\alias{montar_tabela_rendimento_uc}
\title{Montar tabela com rendimento mensal de cada UC (por tipo de rendimento)}
\usage{
montar_tabela_rendimento_uc(
  df_pof_rendimento_base,
  tipo_rendimento = 0,
  df_pof_morador_base
)
}
\arguments{
\item{df_pof_rendimento_base}{O nome (string) do dataframe com os dados de rendimento. Ver ler_pof_rendimento}

\item{df_pof_morador_base}{O nome (string) do dataframe com o registro MORADOR. Ver ler_pof_geral}

\item{tipo_rendimento=0}{Tipo (ou tipos) de rendimento. Ver indice_rendimento}
}
\value{
Um dataframe com todas as UC (registro MORADOR) com os valores de cada tipo de rendimento como colunas
}
\description{
Montar tabela com rendimento mensal de cada UC (por tipo de rendimento)
}
\examples{
montar_tabela_rendimento_uc(df_pof_rendimento_base = "pof_rendimento",
tipo_rendimento = c(0,1,2),
df_pof_morador_base = "pof_morador");
}
\seealso{
ler_pof_rendimento, ler_pof_rendimento_todas, calcular_rendimento_media
}
