% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pof2017package-package.R
\docType{package}
\name{pof2017package-package}
\alias{pof2017package}
\alias{pof2017package-package}
\title{pof2017package: POF (IBGE, Brazil) Data Organized}
\description{
This package aims to provide summary statistics and family-specific data concerning expenditure and income, available at the Brazilian Household Budget Survey (POF - IBGE).
}
\author{
\strong{Maintainer}: Pedro Rubin \email{prubincosta@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
