% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_pof.R
\name{baixar_pof}
\alias{baixar_pof}
\title{Download and unzip POF microdata}
\usage{
baixar_pof(destination_path)
}
\arguments{
\item{destination_path}{The folder in which the microdata files are to be stored (if folder does not exist, it will be created)}
}
\value{
The microdata files in the designated path
}
\description{
Download and unzip POF microdata
}
\examples{
baixar_pof("./microdata");
}
