% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_valor_rendimento_mensal_uc.R
\name{calcular_valor_rendimento_mensal_uc}
\alias{calcular_valor_rendimento_mensal_uc}
\title{Mean monthly income values (by type of income)}
\usage{
calcular_valor_rendimento_mensal_uc(
  pof_rendimento = 0,
  tipo_rendimento = 0,
  uf = "all",
  regiao = "all"
)
}
\arguments{
\item{pof_rendimento}{The name of the df with the income data (string)}

\item{tipo_rendimento=0}{The type (or types) of income. Default to total income.}

\item{path_microdata}{The path to the microdata folder}

\item{uf="all"}{The relevant federal unit (numeric)}

\item{regiao="all"}{The relevant macroregion (character code)}
}
\value{
The mean deflated monthly income
}
\description{
Mean monthly income values (by type of income)
}
\examples{
calcular_valor_rendimento_mensal_uc(tipo_despesa = c(0,1,2), path_midrodata = "./microdata_folder");
calcular_valor_rendimento_mensal_uc(pof_rendimento = "df_income", tipo_despesa = c(0,1,2), regiao = "SE");
}
