% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_pof_rendimento.R
\name{ler_pof_rendimento}
\alias{ler_pof_rendimento}
\title{Carregar microdados da POF - todas as informações de rendimento}
\usage{
ler_pof_rendimento(arquivo_microdados)
}
\arguments{
\item{arquivo_microdados}{O caminho até o arquivo txt do registro de rendimento a ser carregado}
}
\value{
Um dataframe com as informações de rendimento do registro
}
\description{
Carregar microdados da POF - todos os registros com dados de rendimento
Aluguel estimado, caderneta coletiva, despesa coletiva, individual, outros rendimentos e rendimento trabalho
OBS: É preciso que o nome dos arquivos .txt esteja igual a como foi baixado!!
}
\examples{
ler_pof_rendimento("./microdados/ALUGUEL_ESTIMADO.txt");
}
\seealso{
ler_pof_rendimento_todas, baixar_pof
}
