% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_pof_rendimento_todas.R
\name{ler_pof_rendimento_todas}
\alias{ler_pof_rendimento_todas}
\title{Carregar microdados da POF - todas as informações de rendimento}
\usage{
ler_pof_rendimento_todas(pasta_microdados)
}
\arguments{
\item{pasta_microdados}{O caminho para a pasta que contém os arquivos .txt de microdados}
}
\value{
Um dataframe com as informações de rendimento
}
\description{
Carregar microdados da POF - todos os registros com dados de rendimento
Aluguel estimado, caderneta coletiva, despesa coletiva, despesa individual, outros rendimentos e rendimento trabalho
OBS: É preciso que o nome dos arquivos .txt esteja igual a como foi baixado!!
}
\examples{
ler_pof_rendimento_todas("./pasta_microdados");
}
\seealso{
ler_pof_rendimento, baixar_pof
}
