% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montar_tabela_despesa_uc.R
\name{montar_tabela_despesa_uc}
\alias{montar_tabela_despesa_uc}
\title{UC's monthly expenditure values (by type of expenditure)}
\usage{
montar_tabela_despesa_uc(
  pof_despesa,
  tipo_despesa = 0,
  pof_morador,
  uf = "all",
  regiao = "all"
)
}
\arguments{
\item{pof_despesa}{The name of the df with the expenditure data (string). See ler_pof_despesa}

\item{pof_morador}{The path to MORADOR.txt}

\item{tipo_despesa=0}{The type (or types) of expenditure. Default to total expenditure. See indice_despesa}

\item{uf="all"}{The relevant federal unit (numeric). NOT IMPLEMENTED YET}

\item{regiao="all"}{The relevant macroregion (character code). NOT IMPLEMENTED YET}
}
\value{
A datafram with all UC, with the relevant expenditure values as columns
}
\description{
UC's monthly expenditure values (by type of expenditure)
}
\examples{
montar_tabela_despesa_uc(pof_despesa = "df_expenditure", tipo_despesa = c(0,1,2));
}
