% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montar_tabela_rendimento_uc.R
\name{montar_tabela_rendimento_uc}
\alias{montar_tabela_rendimento_uc}
\title{UC's monthly income values (by type of income)}
\usage{
montar_tabela_rendimento_uc(
  pof_rendimento = 0,
  tipo_rendimento = 0,
  path_microdata,
  uf = "all",
  regiao = "all"
)
}
\arguments{
\item{pof_rendimento}{The name of the df with the income data (string).}

\item{path_microdata}{The path to the microdata folder}

\item{tipo_rendimento=0}{The type (or types) of income. Default to total income.}

\item{uf="all"}{The relevant federal unit (numeric)}

\item{regiao="all"}{The relevant macroregion (character code)}
}
\value{
A datafram with all UC, with the relevant income values as columns
}
\description{
UC's monthly income values (by type of income)
}
\examples{
montar_tabela_rendimento_uc(tipo_rendimento = c(0,1,2));
montar_tabela_rendimento_uc(pof_rendimento = "df_income", tipo_rendimento = c(0,1,2), regiao = "N");
}
