% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pof2017package-package.R
\docType{package}
\name{pof2017package-package}
\alias{pof2017package}
\alias{pof2017package-package}
\title{pof2017package: Dados da POF 2017-2018 (IBGE) organizados}
\description{
Esse pacote tem como objetivos facilitar o acesso e análise da POF. Contém funções para baixar os dados, encontrar médias de despesa e rendimento e montar tabelas com dados de despesa e rendimento para cada unidade de consumo.
}
\author{
\strong{Maintainer}: Pedro Rubin \email{prubincosta@gmail.com}

}
\keyword{internal}
